% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computecal.R
\name{computecal}
\alias{computecal}
\title{Prepares a calendar frame}
\usage{
computecal(fy, ly)
}
\arguments{
\item{fy}{first year to work with (past)}

\item{ly}{last year to work with (present)}
}
\value{
3 columns containing year,month,day
}
\description{
This function prepares a calendar frame and returns it as year,month,day (i.e., the 3 first columns of RClimdex format)
}
\examples{
fy<-1981
ly<-2020
clndr<-computecal(fy,ly)
}
