% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dostats.R
\name{dostats}
\alias{dostats}
\title{Create QC statistical summary}
\usage{
dostats()
}
\value{
files with QC summary
}
\description{
This function creates two report files (Mystats.txt and CasesSummary.txt) with a statistical summary
of QCs performed over the whole data set
}
