% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplas.R
\name{duplas}
\alias{duplas}
\title{Detects duplicated dates}
\usage{
duplas(x)
}
\arguments{
\item{x}{a vector of dates in the ECA&D format (YYYYMMDD)}
}
\value{
a vector with the list of positions which do not pass this QC test. If all positions pass the test, returns NULL
}
\description{
This function detects duplicated dates in the input time series
}
