% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flat.R
\name{flat}
\alias{flat}
\title{Flat sequences}
\usage{
flat(y, maxseq, exclude = NULL)
}
\arguments{
\item{y}{a data vector}

\item{maxseq}{the maximum number of contiguous repetitions of a value (e.g., if 3, sequences of 4 will be flagged)}

\item{exclude}{values to be excluded. This is useful for variables where a single value is expected to repeat many times, e.g. 0.0 in precip.}
}
\value{
vector with the list of positions which do not pass this test. If all positions pass the test, returns NULL
}
\description{
This function detects consecutive equal values (e.g., 15.1, 15.1, 15.1, 15.1...). Also can be used to detect consecutive equal decimal part of the values (e.g., 15.1, 12.1, 13.1, 10.1 ...)
}
\examples{
y<-rnorm(100)
y[10:20]<-10
flat(y,5)
}
