% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inqc.R
\name{inqc}
\alias{inqc}
\title{Wraper for QC'ing all varibales}
\usage{
inqc(homefolder = "./", blend = TRUE)
}
\arguments{
\item{homefolder}{a path to the homefolder, as string}

\item{blend}{a logical flag which means performing (if TRUE) QC on blended time series}
}
\value{
QC results, in both formats (verbose and workable file in exact ECA&D format)
}
\description{
This function calls functions which perform QC for all climate variables
}
