% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IQRoutliers.R
\name{IQRoutliers}
\alias{IQRoutliers}
\title{Computes outliers}
\usage{
IQRoutliers(date, value, level = 3, window = 11, exclude = NULL)
}
\arguments{
\item{date}{a vector with dates}

\item{value}{a vector with data values}

\item{level}{a number of IQRs}

\item{window}{a number of days to be considered (including the target)}

\item{exclude}{if it is not null, the code will exlcude this value from the analyisis (i.e., good to exclude 0 for precip)}
}
\value{
positions which do not pass this test (bad)
}
\description{
This function computes outliers centralized around a day, using a number of days around it
}
