% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumps2.R
\name{jumps2}
\alias{jumps2}
\title{Labels interdiurnal large differences}
\usage{
jumps2(date, value, quanty = 0.999, times = 1, force = NULL)
}
\arguments{
\item{date}{vector of dates}

\item{value}{vector of values}

\item{quanty}{numeric value (quantile of the differences) used to aggregate daily differences}

\item{times}{multiplier for daily values differences threshold}

\item{force}{value of threshold for daily values differences to be forced}
}
\value{
both values involved in the jump are returned. Need an additional function to decide which is the culprit
}
\description{
This function labels interdiurnal differences considered too large
}
