% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listas.R
\name{listas}
\alias{listas}
\title{Creates listings for stations linking STAID and SOUID}
\usage{
listas(country = "all", name = "allstations.txt")
}
\arguments{
\item{country}{country for which the list is created. If all, no country filter.}

\item{name}{output file name, do not touch, default is always good.}
}
\value{
a data frame containing all the stations for all elements, linking STAID and SOUID and metadata (a file named as specified by name)
}
\description{
This function takes all the elements and rbinds them into a single list to process
}
