% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listonator.R
\name{listonator}
\alias{listonator}
\title{Creates a list (as 'Global' variable) of stations to be QCed.}
\usage{
listonator(check = TRUE)
}
\arguments{
\item{check}{a logical parameter TRUE/FALSE. If check=TRUE a list of stations is created.}
}
\value{
a list of stations to be QCed (liston)
}
\description{
This function creates a list (and makes it 'Global' variable) of stations to be QCed.
It can be 'blended' or 'non-blanded' stations. Geographical cordinates are transphormed into decimal degrees
}
