% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newfriki.R
\name{newfriki}
\alias{newfriki}
\title{Isolates values which are not continuous in the distribution}
\usage{
newfriki(date, value, margina = 0.999, times = 2)
}
\arguments{
\item{date}{a vector of dates with the format of yyyymmdd}

\item{value}{a vector of data values}

\item{margina}{a tolerence margin, expressed as quantile of the differences}

\item{times}{a multiplicator to the tolerance margin. Intended usage is to run this twice. Once with times = 1 and flag values as suspect; once with times = 2 and flag as error}
}
\value{
positions which do not pass this QC test
}
\description{
The function isolates values which are not continuous in the distribution. If the gap is larger than a preset big margin,
the values above (or below) are flagged
}
