% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretogadget.R
\name{paretogadget}
\alias{paretogadget}
\title{Finds outliyers}
\usage{
paretogadget(x, ret)
}
\arguments{
\item{x}{a vector of values (a series) to be analyzed}

\item{ret}{a pseudo return period}
}
\value{
a list of positions where corresponding values do not pass this QC test (which can be considered as outliyers)
}
\description{
This function finds outliyers for variables which can be described/evaluated by means of the Pareto distribution (e.g. atmospheric precipitation or wind speed)
}
