% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physics.R
\name{physics}
\alias{physics}
\title{Isolates anomalous values}
\usage{
physics(x, nyu = 0, compare = 1)
}
\arguments{
\item{x}{data vector}

\item{nyu}{threshold, numeric}

\item{compare}{logical operation to apply over the threshold. 1: larger; 2: larger or equal; 3: smaller; 4: smaller or equal; 5 equal}
}
\value{
list of positions which do not pass this QC test. If all positions pass the test, returns NULL
}
\description{
Given a data vector, the function will compare the values to the specified threshold
}
\examples{
x<-rnorm(100)
x[10]<-100
physics(x,5,1)
}
