% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potpareto.R
\name{potpareto}
\alias{potpareto}
\title{Peaks over threshold modelling}
\usage{
potpareto(y, thres = 0.99)
}
\arguments{
\item{y}{a vector of values (a series) to be analyzed}

\item{thres}{a threshold value of probability to define a coresponding threshold percentile}
}
\value{
a list containing results of modelling/fitting the generalized Pareto distribution
}
\description{
This function fits the Generalized Pareto distrebution for exeedances over a threshold
}
