% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precip.R
\name{precip}
\alias{precip}
\title{QC for Atmospheric Precipitation (RR)}
\usage{
precip(
  large = 5000,
  small = 0,
  element = "RR",
  ret = 500,
  retornoracha = 500,
  margin = 20,
  friki = 150,
  blocksizeround = 20,
  blockmanymonth = 15,
  blockmanyyear = 180,
  limit = 1500,
  tolerance = 8,
  maxseq = 3,
  roundmax = 10,
  level = 15,
  window = 30,
  margina = 0.999,
  inisia = FALSE
)
}
\arguments{
\item{large}{value above which the observation is considered physically impossible for the region}

\item{small}{value below which the observation is considered physically impossible for the region}

\item{element}{two-letters ECA&D code for the element (RR for precipitation)}

\item{ret}{pseudo-return period for the pareto outliers}

\item{retornoracha}{return period for the calculation of the maximum dry and wet spell}

\item{margin}{frequency difference between consecutive values for repeatedvalue()}

\item{friki}{minimum value to be considered by repeatedvalue()}

\item{blocksizeround}{the maximum number of repeated values, FUNCTION roundprecip}

\item{blockmanymonth}{maximum number of equal values in a month, FUNCTION: toomany}

\item{blockmanyyear}{maximum number of equal values in a yaer, FUCNTION: toomany}

\item{limit}{cut threshold for FUNCTION suspectacumprec}

\item{tolerance}{number of NA or 0s before allowed before the limit, FUNCTION suspectacumprec}

\item{maxseq}{maximum number of consecutive repeated values, FUNCTION: flat  (11.1,11.1,11.1 would be 3 consecutives)}

\item{roundmax}{maximum number of consecutive decimal part value, for flat function (10.0, 11.0, 12.0 would be 3 consecutive)}

\item{level}{level for IQRoutliers}

\item{window}{window for IQRoutliers}

\item{margina}{should be described later}

\item{inisia}{a logical flag}
}
\value{
results of QC for RR
}
\description{
This function will centralize precipitation-like qc routines. Will create a file in the folder QC
with an additional 0/1 column, where "1" means test failed.
}
