% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/putjulian.R
\name{putjulian}
\alias{putjulian}
\title{Merge julian days}
\usage{
putjulian(x)
}
\arguments{
\item{x}{nxn with year,month,day and data columns}
}
\value{
nxn+1 columns with year,month,day, julian and data
}
\description{
This function merges julian days to a yyyy,mm,dd and data
}
