% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relhum.R
\name{relhum}
\alias{relhum}
\title{QC for Relative Humidity (HU)}
\usage{
relhum(
  element = "HU",
  maxseq = 3,
  blocksizeround = 20,
  blockmanymonth = 15,
  blockmanyyear = 180,
  roundmax = 10,
  inisia = FALSE
)
}
\arguments{
\item{element}{two-letters ECA&D code for the element (HU for relative humidity)}

\item{maxseq}{maximum number of consecutive repeated values, for flat function (11.1,11.1,11.1 would be 3 consecutives). Passed on to flat(). See ?flat for details}

\item{blocksizeround}{maximum number of values in a month with the same decimal, FUNCTION: rounding}

\item{blockmanymonth}{maximum number of equal values in a month, FUNCTION: toomany}

\item{blockmanyyear}{maximum number of equal values in a yaer, FUCNTION: toomany}

\item{roundmax}{maximum number of consecutive decimal part value, for flat function (10.0, 11.0, 12.0 would be 3 consecutive)}

\item{inisia}{a logical flag. If it is TRUE inithome() will be called}
}
\value{
results of QC for HU
}
\description{
This function will centralize temperature-like qc routines. Will create a file in the folder QC
with an additional 0/1 column, where "1" means test failed.
}
