% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeatedvalue.R
\name{repeatedvalue}
\alias{repeatedvalue}
\title{Find repeated values}
\usage{
repeatedvalue(x, margin = 20, friki = 150)
}
\arguments{
\item{x}{preciptiation time series}

\item{margin}{the difference in frequency of the nearest value}

\item{friki}{the minimum value to be considered}
}
\value{
list of values which do not pass QC (target)
}
\description{
This function looks for a value which repeats too many times and, given the decaying shape of preciptiation empirical distro,
is considered too large to happen that many times
}
