% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnpotpareto.R
\name{returnpotpareto}
\alias{returnpotpareto}
\title{It should be described later}
\usage{
returnpotpareto(pato, ret, w = 1.65)
}
\arguments{
\item{pato}{it should be described later}

\item{ret}{it should be described later}

\item{w}{it should be described later}
}
\value{
list of values which did not pass the test
}
\description{
This function should be described later
}
