% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selepe.R
\name{selepe}
\alias{selepe}
\title{QC for Atmospheric Pressure (PP)}
\usage{
selepe(
  large = 15000,
  small = 8000,
  maxjump = 2000,
  maxseq = 3,
  margina = 0.999,
  level = 5,
  window = 30,
  roundmax = 10,
  blocksize = 10,
  step = 30,
  blockmanymonth = 15,
  blockmanyyear = 180,
  blocksizeround = 20,
  qjump = 0.999,
  tjump = 1.5,
  element = "PP",
  inisia = FALSE
)
}
\arguments{
\item{large}{value above which the observation is considered physically impossible for the region}

\item{small}{value below which the observation is considered physically impossible for the region}

\item{maxjump}{it should be described later}

\item{maxseq}{maximum number of consecutive repeated values, for flat function (11.1,11.1,11.1 would be 3 consecutives)}

\item{margina}{quantile for newfriki function}

\item{level}{number of IQRs for IQR outliers}

\item{window}{window, in days, for IQR outliers}

\item{roundmax}{maximum number of consecutive decimal part value, for flat function (10.0, 11.0, 12.0 would be 3 consecutive)}

\item{blocksize}{should be described later}

\item{step}{should be described later}

\item{blockmanymonth}{maximum number of equal values in a month, FUNCTION: toomany}

\item{blockmanyyear}{maximum number of equal values in a yaer, FUCNTION: toomany}

\item{blocksizeround}{maximum number of values in a month with the same decimal, for rounding function}

\item{qjump}{quantile for the calculated maximum jump allowed}

\item{tjump}{factor to multiply the qjump computed differences}

\item{element}{two-letters ECA&D code for the element (PP for sea level pressure)}

\item{inisia}{a logical flag. If it is TRUE inithome() will be called}
}
\value{
results of QC for PP
}
\description{
This function will centralize temperature-like qc routines. It will create a file in the folder QC
with an additional 0/1 column where "1" means test failed.
}
