% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snowdepth.R
\name{snowdepth}
\alias{snowdepth}
\title{QC for Snow Depth (SD)}
\usage{
snowdepth(
  element = "SD",
  maxseq = 20,
  blocksizeround = 20,
  blockmanymonth = 20,
  blockmanyyear = 200,
  large = 5000,
  exclude = 0,
  inisia = FALSE
)
}
\arguments{
\item{element}{two-letters ECA&D code for the element (SD for snow depth)}

\item{maxseq}{maximum number of consecutive repeated values, FUNCTION: flat() (11.1,11.1,11.1 would be 3 consecutives)}

\item{blocksizeround}{maximum number of values in a month with the same decimal, FUNCTION: rounding()}

\item{blockmanymonth}{maximum number of equal values in a month, FUNCTION: toomany()}

\item{blockmanyyear}{maximum number of equal values in a yaer, FUCNTION: toomany()}

\item{large}{value above which the observation is considered physically impossible for the region, FUNCTION: physics()}

\item{exclude}{value to be excluded from a function (in this case, 0 for flats)}

\item{inisia}{a logical flag. If it is TRUE inithome() will be called}
}
\value{
results of QC for SD
}
\description{
This function will centralize temperature-like qc routines. It will create a file in the folder QC
with an additional 0/1 column, where "1" means test failed.
}
