% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suspectacumprec.R
\name{suspectacumprec}
\alias{suspectacumprec}
\title{Detect precipitation values above limit}
\usage{
suspectacumprec(datos, limit = 2000, tolerance = 10)
}
\arguments{
\item{datos}{a two columns vector, date and data, in the ECA&D format}

\item{limit}{el valor de corte para estudiar la precipitacion}

\item{tolerance}{cuantos días consecutivos con 0 o NA necesita para saltar}
}
\value{
el indice de los días que no pasan el qc($chungos) and la serie rellenada con computecal (es decir, sin huecos) ($px)
}
\description{
This function detects values above limit preceded by a number of "non precip days", given by tolerance
pretende la detección de precipitación acumulada
}
