% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toomany.R
\name{toomany}
\alias{toomany}
\title{Looks if a value is repeated too many times}
\usage{
toomany(y, blockmany = 15, scope = 1, exclude = NULL)
}
\arguments{
\item{y}{two columns with date and data}

\item{blockmany}{maximum number of repeated values in a month, year, or season}

\item{scope}{monthly (1), annual (2)}

\item{exclude}{values to exclude, e.g. if precip, 0 must be excluded}
}
\value{
list of positions which do not pass this QC test. If all positions pass the test, returns NULL
}
\description{
This function splits data by month and looks if a value is repeated too many times
}
