% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inithome.R
\name{inithome}
\alias{inithome}
\title{Creates necessary folders (if not exist)}
\usage{
inithome()
}
\value{
it does not return any values, just creates the described folders if they do not exist
}
\description{
This function will checks if all necessary folders ('QCSummary, QC and QCConsolidated) exist and if not, creates them.
Not intended as a stand-alone function. Called from other routines.
}
