% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumps2.R
\name{jumps2}
\alias{jumps2}
\title{Labels interdiurnal differences considered too large}
\usage{
jumps2(date, value, quanty = 0.999, times = 1, force = NULL)
}
\arguments{
\item{date}{it should be described later}

\item{value}{vector of values}

\item{quanty}{it should be described later}

\item{times}{it should be described later}

\item{force}{it should be described later}
}
\value{
chungo both values involved in the jump are returned. Need an additional function to decide which is the culprit
}
\description{
The function labels interdiurnal differences considered too large
}
