% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lister.R
\name{lister}
\alias{lister}
\title{Creates a list of blended/non-bladed files for some climate variable}
\usage{
lister(element)
}
\arguments{
\item{element}{a climatological element (defined by means of two letters, i.e. 'TX')}
}
\value{
a list of blended or non-bladed files to be QCed
}
\description{
This function create a list of blended or non-bladed files contaning data of a specified element to be QCed.
}
