% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paretogadget.R
\name{paretogadget}
\alias{paretogadget}
\title{It should be described later}
\usage{
paretogadget(x, ret)
}
\arguments{
\item{x}{the series you are going to use (will take precip)}

\item{ret}{the pseudo return period (remember is pareto and I destroyed the years)}
}
\value{
list of values which did not pass the test
}
\description{
This function should be described later
}
