% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potpareto.R
\name{potpareto}
\alias{potpareto}
\title{It should be described later}
\usage{
potpareto(y, thres = 0.99)
}
\arguments{
\item{y}{it should be described later}

\item{thres}{it should be described later}
}
\value{
list of values which did not pass the test
}
\description{
This function should be described later
}
