% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readheader.R
\name{readheader}
\alias{readheader}
\title{Read the header of an ECA&D file}
\usage{
readheader(input = "SS_STAID000143.txt")
}
\arguments{
\item{input}{an ECA&D filename}
}
\value{
The header of an ECA&D file
}
\description{
Reads one ECA&D file and returns the header, so it can be written in the same way
}
