% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeatedvalue.R
\name{repeatedvalue}
\alias{repeatedvalue}
\title{Finds repeated values}
\usage{
repeatedvalue(x, margin = 20, friki = 150)
}
\arguments{
\item{x}{precipitation time series}

\item{margin}{difference in frequency of the nearest value}

\item{friki}{minimum value to be considered}
}
\value{
list of positions which do not pass this QC test. If all positions pass the test, returns NULL
}
\description{
This function looks for a value which repeats too many times and, given the decaying shape of precipitation of empirical distribution,
is considered too large to happen that many times
}
