% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnpotpareto.R
\name{returnpotpareto}
\alias{returnpotpareto}
\title{Threshold percentile for the Pareto outliers}
\usage{
returnpotpareto(pato, ret, w = 1.65)
}
\arguments{
\item{pato}{list with results of modelling/fitting the generalized Pareto distribution}

\item{ret}{pseudo return period}

\item{w}{multiplicator used to calculate a threshold probability}
}
\value{
percentile of the generalized Pareto distribution corresponding to a threshold probability calculated based on the peudo return period
}
\description{
This function returns a value of a threshold percentile for the Pareto outliyers
}
