% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rounding.R
\name{rounding}
\alias{rounding}
\title{Detects rounded sections}
\usage{
rounding(y, blocksize = 20)
}
\arguments{
\item{y}{two columns with date in the ECA&D format (yyyymmdd) and data}

\item{blocksize}{maximum number of repeated values with the same decimal allowed on each block (blocks = months)}
}
\value{
list of positions which do not pass this QC test. If all positions pass the test, returns NULL
}
\description{
This function splits data by month and looks if a decimal value is repeated too many times
}
