% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundprecip.R
\name{roundprecip}
\alias{roundprecip}
\title{Rounding in precipitation data}
\usage{
roundprecip(y, blocksize = 20, exclude = 0)
}
\arguments{
\item{y}{two columns with date and data}

\item{blocksize}{maximum number of repeated values with the same decimal}

\item{exclude}{value to be excluded (zero for precipitation)}
}
\value{
list of positions which do not pass this QC test. If all positions pass the test, returns NULL
}
\description{
This function splits data by month and looks if a decimal value is repeated too many times
}
